move plugin helpers to own directory
reason: the plugin directory was not the right place
