+++ wazo_phoned/plugin_helpers/proxy/resource.py
