+++ wazo_phoned/plugins/thomson/resource.py
@@ -1,7 +1,7 @@
 # Copyright 2019 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from wazo_phoned.plugins.proxy_helpers.resource import (
+from wazo_phoned.plugin_helpers.proxy.resource import (
     ProxyLookup,
 )
